chore: split to dev and prod folders
