+++ packages/devextreme/js/__internal/core/state_manager/dev/setup_state_manager.ts
@@ -1,5 +1,5 @@
 /* eslint-disable spellcheck/spell-checker */
-import type { DIContext } from '../di';
+import type { DIContext } from '../../di';
 import { Logger } from './logger';
 import { StateManagerFactory } from './state_manager';
 import type * as StateManagementTypes from './types';
