+++ packages/devextreme/js/__internal/core/state_manager/dev/types.ts
