+++ packages/devextreme/js/__internal/core/state_manager/index.ts
@@ -1,13 +1,1 @@
-export type {
-  ReadonlySignal,
-  Signal,
-} from './reactive_primitives/index';
-export {
-  batch,
-  computed,
-  effect,
-  signal,
-  // eslint-disable-next-line spellcheck/spell-checker
-  untracked,
-} from './reactive_primitives/index';
-export { setupStateManager } from './setup_state_manager';
+export * from './dev/index';
