+++ packages/devextreme/js/__internal/core/state_manager/prod/index.ts
